+++ component/backend/View/hhvm.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @package   AkeebaBackup
+ * @package   AkeebaLoginGuard
  * @copyright Copyright (c)2016-2018 Nicholas K. Dionysopoulos / Akeeba Ltd
  * @license   GNU General Public License version 3, or later
  */