+++ component/backend/loginguard.php
@@ -41,6 +41,14 @@ if (defined('HHVM_VERSION'))
 	return;
 }
 
+// So, FEF is not installed?
+if (!@file_exists(JPATH_SITE . '/media/fef/fef.php'))
+{
+	(include_once __DIR__ . '/View/fef.php') or die('You need to have the Akeeba Frontend Framework (FEF) package installed on your site to display this component. Please visit https://www.akeeba.com/download/official/fef.html to download it and install it on your site.');
+
+	return;
+}
+
 
 /**
  * The following code is a neat trick to help us collect the maximum amount of relevant information when a user
@@ -62,7 +70,9 @@ function mainLoopAkeebaLoginGuard()
 {
 	if (!defined('FOF30_INCLUDED') && !@include_once(JPATH_LIBRARIES . '/fof30/include.php'))
 	{
-		throw new RuntimeException('FOF 3.0 is not installed', 500);
+		(include_once __DIR__ . '/View/fof.php') or die('You need to have the Akeeba Framework-on-Framework (FOF) 3 package installed on your site to use this component. Please visit https://www.akeeba.com/download/fof3.html to download it and install it on your site.');
+
+		return;
 	}
 
 	FOF30\Container\Container::getInstance('com_loginguard')->dispatcher->dispatch();