+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # 2.0.0
 
-**Other changes**
+**New**
 
 * Rewritten interface using the FOF framework
 * Rewritten interface using our own CSS framework
 * Preliminary Joomla! 4 compatibility (tested against 4.0.0 Alpha 2)
 * Minimum requirements increased to PHP 5.4 or later. Tested up to and including PHP 7.2.
+* Warn the user if either FOF or FEF is not installed.
 
 **Bug fixes**